Modified @Extension annotations to make EnvInject ordered
The default order used by Jenkins is unpredictable and for
my use case I wanted EnjInject to be after other BuildWrappers
but before the Artifactory plugin build wrappers. So, this
simple changes makes sure that unless the order is Groovy
adjusted the default order makes this the first of the last
group of plugins. Ordinal -1